examples/langchain/requirements.txt
streamlit
pandas
python-dotenv
langchain
pathlib
PyJWT
openai
tiktoken
faiss-cpu